requirements.txt
flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
langchain-openai==0.1.8
langchain==0.1.9
certifi==2024.2.2
flask-cors==4.0.0
gunicorn==21.2.0 